Prise en compte des remarques de la PR #64
