+++ ci-travis/scripts/echoColor.sh
@@ -17,22 +17,22 @@ function echoC
 	case "$color" in
 	"r" | "red" | "RED")
 	    START="\e[31m"
-	    ;
+	    ;;
 	"g" | "green" | "GREEN")
 	    START="\e[32m"
-	    ;
+	    ;;
 	"o" | "orange" | "ORANGE")
 	    START="\e[33m"
-	    ;
+	    ;;
 	"b" | "blue" | "BLUE")
 	    START="\e[34m"
-	    ;
+	    ;;
 	"p" | "purple" | "PURPLE")
 	    START="\e[35m"
-	    ;
+	    ;;
 	"lb" | "lightblue" | "LIGHTBLUE")
 	    START="\e[36m"
-	    ;
+	    ;;
 	*)
 		START=$END
 		;
